requirements.txt
# Deep Learning Frameworks
torch>=2.0,<3.0
torchvision>=0.15,<1.0

# Computer Vision & Image Processing
opencv-python>=4.8,<5.0
Pillow>=10.0,<11.0
albumentations>=1.3,<2.0

# Machine Learning & Data Science
scikit-learn>=1.3,<2.0
pandas>=2.0,<3.0
numpy>=1.24,<2.0

# Visualization
matplotlib>=3.7,<4.0
seaborn>=0.12,<1.0
plotly>=5.14,<6.0

# Web Application
streamlit>=1.28,<2.0

# Explainability
grad-cam>=1.4,<2.0

# Utilities
tqdm>=4.65,<5.0
openpyxl>=3.1,<4.0
